+++ databento/common/types.py
@@ -1,4 +1,4 @@
-from typing import Callable, Union
+from typing import Callable, Generic, TypeVar, Union
 
 import databento_dbn
 
@@ -21,3 +21,34 @@ DBNRecord = Union[
 
 RecordCallback = Callable[[DBNRecord], None]
 ExceptionCallback = Callable[[Exception], None]
+
+_T = TypeVar("_T")
+class Default(Generic[_T]):
+    """
+    A container for a default value. This is to be used when a callable wants
+    to detect if a default parameter value is being used.
+
+    Example
+    -------
+        def foo(param=Default[int](10)):
+            if isinstance(param, Default):
+                print(f"param={param.value} (default)")
+            else:
+                print(f"param={param.value}")
+
+    """
+
+    def __init__(self, value: _T):
+        self._value = value
+
+    @property
+    def value(self) -> _T:
+        """
+        The default value.
+
+        Returns
+        -------
+        _T
+
+        """
+        return self._value